AplikacjaDjango/requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==39.0.0
Django==4.1.3
django-classy-tags==4.0.0
django-cms==3.11.1
django-cors-headers==3.13.0
django-extensions==3.2.1
django-filter==22.1
django-formtools==2.4
django-rest-framework==0.1.0
django-rest-knox==4.2.0
django-sekizai==4.0.0
django-treebeard==4.4
djangocms-admin-style==3.2.1
djangocms-snippet==3.1.0
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
djangorestframework-jwt==1.11.0
inflection==0.5.1
Markdown==3.4.1
psycopg2==2.8.6
pycparser==2.21
PyJWT==1.7.1
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.6
